server/requirements.txt
cryptography==41.0.3
bcrypt==3.2.2
PyJWT==2.8.0